Rename files to match naming scheme for 2.3.0.0
In version 2.3.0.0 of OggDude's character generator, filenames usually
have spaces in them rather than underscores.

A couple of files were deleted rather than renamed, because they were
duplicates: "Unmatched_Survivability.xml" had a duplicate file in the
misspelled "Umatched_Survivability.xml", and the misspelled file was
better. So I kept the better file and deleted the misspelling.

The other file that was deleted rather than renamed was the file named
"AD-15_Modular_Multi-Role_Starfighter.xml". This was a misspelling of
the "AD-1S" vehicle, so I kept the "AD-1S" file and deleted "AD-15".