Import data from final BigTent upload
The last file uploaded to the BigTent group before it shut down was
"Full Dataset for PreRelease 2.3.0.0", created by lee_underdunk on
September 5, 2018. Its description on BigTent was: "Added descriptions
for Fully Operational, and everything I could for Ghosts of Dathomir,
Dawn of Rebellion, Unlimited Power, Knights of Fate, and Cyphers and
Masks. A few things are a bit of kludge to make them work but hopefully
most of it is correct."

This greatly increases the number of descriptions available.