Update all but Vehicles folder to SWCharGen 2.3.4.0
This commit consists of importing the unmodified data from version
2.3.4.0 of OggDude's generator, then copying all the descriptions from
the SWDataSet files back into the folder. This should eliminate many of
the problems people were encountering with duplicate items caused by
item keys being different between the SWDataSet version of an item and
the OggDude version. E.g., the RLR-331 "Bulldog" Rocket Rifle from
Knights of Fate was added manually to the SWDataSet at one point when
OggDude's generator did not yet have any items from Knights of Fate. At
the time, the person who added it used the key RLR331BULLDOG. Later,
OggDude added the Knights of Fate items, and chose the key RLR331 for
the RLR-331 rifle. Since OggDude's generator uses the keys to determine
whether two entries are for the same item or two different items, it
considered them to be two different items even though they were clearly
meant to be the same item. This commit should fix all instances of
duplicate items like this due to different keys (unless I missed one),
except for the Vehicles folder. The Vehicles folder, with its 400 or so
data files, will be updated in a different commit.