Updated vehicles to 2.3.4.0
The following files were removed because they were duplicates:

- AD-15 Modular Multi-Role Starfighter.xml
- Kappa-Class Shuttle (Rebellion).xml
- MC40A Light Cruiser (1).xml
- TX-225 GAVw Combat Assault Tank.xml
- TX-225 GAVw Combat AssaultTank.xml
- Tallanx-Class Stealth Fighter (1).xml

The following files were kept because they were not in the OggDude
data (and were not a duplicate of anything else):

- All Terrain Armored Cargo Transport.xml
- All Terrain Defense Pod.xml
- Auzituck-class Gunship.xml
- DS-1 Death Star.xml
- Delta-class T-3c Shuttle.xml
- Ghost.xml
- Hammerhead-class Corvette.xml
- KST-100 Kestrel Light Executive Transport.xml
- Phantom.xml
- Phoenix Cell's Cargo Skiff.xml
- Prototype-B6.xml
- Sabine's TIE Fighter.xml
- Shadow Caster.xml
- TIE Advanced V1.xml
- TIE_MG Mining Guild Starfighter.xml
- TIE_SK Striker Multi-Role Fighter.xml
- TX-225 GAVw _Occupier_ Combat Assault Tank.xml
- The Booster BX.xml
- The Broken Horn.xml
- UT-60D U-Wing Troop Transport.xml
- YV-929 Light Freighter.xml
- Zeta-class Heavy Cargo Shuttle.xml

Finally, the Mandalorian Protectorate _Fang_ Starfighter.xml file was
kept because although it is a near-duplicate of the Fang Fighter.xml
file in the OggDude data, it has slightly different stats, so either
this is a vehicle with multiple variants, or else one of these is a
duplicate that can be deleted later. For now, we'll err on the side of
prudence and keep both files.